Upgrade to latest SoftFloat

+++ riscv/insns/fclass_d.h
@@ -1,3 +1,3 @@
 require_extension('D');
 require_fp;
-WRITE_RD(f64_classify(FRS1));
+WRITE_RD(f64_classify(f64(FRS1)));
