+++ riscv/insns/fcvt_l_d.h
@@ -2,5 +2,5 @@ require_extension('D');
 require_rv64;
 require_fp;
 softfloat_roundingMode = RM;
-WRITE_RD(f64_to_i64(FRS1, RM, true));
+WRITE_RD(f64_to_i64(f64(FRS1), RM, true));
 set_fp_exceptions;
