+++ riscv/insns/feq_d.h
@@ -1,4 +1,4 @@
 require_extension('D');
 require_fp;
-WRITE_RD(f64_eq(FRS1, FRS2));
+WRITE_RD(f64_eq(f64(FRS1), f64(FRS2)));
 set_fp_exceptions;
