+++ riscv/insns/flt_d.h
@@ -1,4 +1,4 @@
 require_extension('D');
 require_fp;
-WRITE_RD(f64_lt(FRS1, FRS2));
+WRITE_RD(f64_lt(f64(FRS1), f64(FRS2)));
 set_fp_exceptions;
