+++ riscv/insns/flt_s.h
@@ -1,4 +1,4 @@
 require_extension('F');
 require_fp;
-WRITE_RD(f32_lt(FRS1, FRS2));
+WRITE_RD(f32_lt(f32(FRS1), f32(FRS2)));
 set_fp_exceptions;
